Remove usage of keys in combiners and reducers
The usage of two keys (stats1 and stats2) gives us better speedups
in practice, but actually takes more time to compute if we compare
to when we use no keys at all, regardless of numbers of cores.

For combiners, using 2 separate keys doesn't even make sense, since
mrjob already parallelize by splitting the data into different cores
anyway. Using the keys here can even bring the overhead of having to
write more data into the disk, since the mappers would have to yield
2 keys for each of the lines.

For reducers, using 2 keys can have a speedup of maximum 2. However,
the workload for the first key (stats1) is too small, so it is not
worth it to use the keys here either.

Generally, it is much better to parallelize the data instead of the
keys. An added benefit is that the code looks much cleaner now.